Fix possible integer overflow
This caused a segfault on a cloud of 2Gb, which should not overflow
uint32's.

+++ io/src/pcd_io.cpp
@@ -602,7 +602,7 @@ pcl::PCDReader::readBodyBinary (const unsigned char *map, pcl::PCLPointCloud2 &c
 
     // Unpack the xxyyzz to xyz
     std::vector<char*> pters (fields.size ());
-    int toff = 0;
+    size_t toff = 0;
     for (size_t i = 0; i < pters.size (); ++i)
     {
       pters[i] = &buf[toff];
